Modularize selection events tracking
